Get historical market data

Retrieves market data for a coin for the last n days

coin_history(
  coin_id,
  vs_currency = "usd",
  days,
  interval = NULL,
  max_attempts = 3
)

Arguments

coin_id
(character): ID of the coin of interest. An up-to-date list of supported coins and their IDs can be retrieved with the supported_coins() function.

vs_currency
(character): name of the reference currency to express the price in. An up-to-date list of supported reference currencies (both fiat and cryptocurrencies) can be obtained with the supported_currencies() function. If an unsupported vs_currency is requested, the call will fail with the respective error message.

days
(numeric or "max"): number of days to look back. If days = "max", the entire available history for coin_id will be retrieved. Depending on the value of days, the time interval used to present the data will differ - see "Details".

interval
(character or NULL): time interval used to present the data. The only currently supported value is daily. Defaults to NULL.

max_attempts
(double, positive): specifies the maximum number of attempts to call the CoinGecko API (e.g., if the first call fails for some reason). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with the following columns:

timestamp (POSIXct);

coin_id (character): same as the argument coin_id;

vs_currency (character): same as the argument vs_currency;

price (double): coin price, as of timestamp;

total_volume (double): a 24 hours rolling-window trading volume, as of timestamp;

market_cap (double): market capitalisation, as of timestamp.

Details

If days = 1 and interval = NULL, the data will be returned for every few minutes (typically 3-8 minutes). If days is between 2 and 90 (inclusive) and interval = NULL, an (approximately) hourly time step will be used. Daily data are used for days above 90. If interval = "daily", daily data will be used irrespective of the value of days.

This function is based on the public CoinGecko API, which has a limit of 50 calls per minute. Please keep this limit in mind when developing your applications.

Examples

# \donttest{
r <- coin_history(coin_id = "bitcoin", vs_currency = "usd", days = 30)
print(r)

#> # A tibble: 721 x 6
#>    timestamp           coin_id vs_currency  price total_volume    market_cap
#>    <dttm>              <chr>   <chr>        <dbl>        <dbl>         <dbl>
#>  1 2021-07-12 11:51:00 bitcoin usd         33705. 19798408238. 632506446187.
#>  2 2021-07-12 11:03:35 bitcoin usd         33747. 19452141518. 632679635740.
#>  3 2021-07-12 10:01:10 bitcoin usd         34185. 19123518019. 641341225415.
#>  4 2021-07-12 09:01:16 bitcoin usd         34323. 19132119089. 643695504310.
#>  5 2021-07-12 08:00:25 bitcoin usd         34357. 19020435107. 645137768238.
#>  6 2021-07-12 07:00:49 bitcoin usd         34437. 18900002368. 645846833924.
#>  7 2021-07-12 06:01:40 bitcoin usd         34257. 18490179943. 641826036668.
#>  8 2021-07-12 05:01:03 bitcoin usd         34371. 18719209402. 645243754813.
#>  9 2021-07-12 04:01:38 bitcoin usd         34597. 18402000751. 646774406080.
#> 10 2021-07-12 03:01:04 bitcoin usd         34298. 17782445967. 643232464120.
#> # … with 711 more rows

# }