Exchange tickers

Retrieves the most recent exchange tickers for a coin

coin_tickers(coin_id, exchange_id, max_attempts = 3)

Arguments

coin_id
(character): ID of the coin of interest. An up-to-date list of supported coins and their IDs can be retrieved with the supported_coins() function.

exchange_id
(character): ID of the exchange to retrieve the data from. An up-to-date list of supported exchanges and their IDs can be obtained with the supported_exchanges() function.

max_attempts
(double, positive): specifies the maximum number of attempts to call the CoinGecko API (e.g., if the first call fails for some reason). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with the following columns:

exchange_id (character): same as the argument exchange_id;

exchange_name (character): common name of the exchange;

coin_id (character): same as the argument coin_id;

name (character): common name of the coin;

base (character): symbol of the base currency in a trading pair;

target (character): symbol of the target currency in a trading pair;

trust_score (character): trust score of the trading pair ("green", "yellow", "red"; see this and this article on the CoinGecko website for details);

last_price (double): last price reported by this exchange for this trading pair;

last_fetch_at (POSIXct, UTC time zone): timestamp of when last_price was recorded;

last_traded_at (POSIXct, UTC time zone): timestamp of the most recent trade;

bid_ask_spread_percentage (double): percentage difference between the ask price (lowest price a seller is willing to sell for) and the bid price (highest price a buyer is willing to buy for; see Investopedia for details);

trading_volume_24h (double): trading volume (in target currency) recorded in the last 24 h (as of last_traded_at);

last_price_btc (double): last price reported by this exchange for this coin, expressed in Bitcoin;

last_price_eth (double): last price reported by this exchange for this coin, expressed in Ethereum;

last_price_usd (double): last price reported by this exchange for this coin, expressed in US dollar;

trading_volume_24h_btc (double): 24 hours trading volume expressed in Bitcoin (as of last_traded_at);

trading_volume_24h_eth (double): 24 hours trading volume expressed in Ethereum;

trading_volume_24h_usd (double): 24 hours trading volume expressed in US dollars;

cost_to_move_up_2percent_usd and cost_to_move_down_2percent_usd (double): 2% market depth (see this article on the CoinGecko website for details);

is_anomaly (logical): an indicator of whether the ticker's price is an outlier (see "Outlier detection" in Methodology on the GoinGecko website);

is_stale (logical): an indicator of whether the ticker's price has not been updated for a while (see "Outlier detection" in Methodology on the GoinGecko website);

trade_url (character): URL to this trading pair's page.

Details

This function is based on the public CoinGecko API, which has a limit of 50 calls per minute. Please keep this limit in mind when developing your applications.

Examples

# \donttest{
r <- coin_tickers(coin_id = "cardano", exchange_id = "binance")
print(r)
#> # A tibble: 14 × 23
#>    exchange_id exchange_name coin_id name    base  target trust_score last_price
#>    <chr>       <chr>         <chr>   <chr>   <chr> <chr>  <chr>            <dbl>
#>  1 binance     Binance       cardano Cardano ADA   USDT   green          1.95e+0
#>  2 binance     Binance       cardano Cardano ADA   BTC    green          3.21e-5
#>  3 binance     Binance       cardano Cardano ADA   BUSD   green          1.95e+0
#>  4 binance     Binance       cardano Cardano ADA   USDC   green          1.96e+0
#>  5 binance     Binance       cardano Cardano ADA   ETH    green          4.58e-4
#>  6 binance     Binance       cardano Cardano ADA   BNB    green          3.76e-3
#>  7 binance     Binance       cardano Cardano ADA   EUR    green          1.70e+0
#>  8 binance     Binance       cardano Cardano ADA   TRY    green          1.88e+1
#>  9 binance     Binance       cardano Cardano ADA   AUD    green          2.61e+0
#> 10 binance     Binance       cardano Cardano ADA   GBP    green          1.42e+0
#> 11 binance     Binance       cardano Cardano ADA   BIDR   green          2.78e+4
#> 12 binance     Binance       cardano Cardano ADA   RUB    green          1.39e+2
#> 13 binance     Binance       cardano Cardano ADA   TUSD   green          1.95e+0
#> 14 binance     Binance       cardano Cardano ADA   BRL    NA             1.12e+1
#> # … with 15 more variables: last_fetch_at <dttm>, last_traded_at <dttm>,
#> #   bid_ask_spread_percentage <dbl>, trading_volume_24h <dbl>,
#> #   last_price_btc <dbl>, last_price_eth <dbl>, last_price_usd <dbl>,
#> #   trading_volume_24h_btc <dbl>, trading_volume_24h_eth <dbl>,
#> #   trading_volume_24h_usd <int>, cost_to_move_up_2percent_usd <dbl>,
#> #   cost_to_move_down_2percent_usd <dbl>, is_anomaly <lgl>, is_stale <lgl>,
#> #   trade_url <chr>
# }