Current coin prices

Retrieves current prices of supported coins in any supported reference currencies

current_price(
  coin_ids,
  vs_currencies = c("usd"),
  include_market_cap = TRUE,
  include_24h_vol = TRUE,
  include_24h_change = TRUE,
  max_attempts = 3
)

Arguments

coin_ids
(character): a vector with IDs of the coins of interest. An up-to-date list of supported coins and their IDs can be obtained with the supported_coins() function.

vs_currencies
(character): a vector with names of the reference currencies to express the price in, e.g. c("usd", "eur", "btc"). An up-to-date list of supported vs_currencies (both fiat and cryptocurrencies) can be obtained with the supported_currencies() function. If vs_currencies contains at least one unsupported currency, the call will fail with the respective error message.

include_market_cap
(boolean, defaults to TRUE): whether to return the market capitalisation information.

include_24h_vol
(boolean, defaults to TRUE): whether to return the trading volume for the last 24 hours.

include_24h_change
(boolean, defaults to TRUE): whether to return the price percentage change compared to 24 hours ago.

max_attempts
(double, positive): specifies the maximum number of attempts to call the CoinGecko API (e.g., if the first call fails for some reason). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble, which by the default will contain the following columns (use arguments include_market_cap, include_24h_vol and include_24h_change to control the inclusion of the corresponding columns):

coin_id (character): coin IDs, ordered alphabetically;

vs_currency (character): reference currency, in which the price of coin_id is expressed;

last_updated_at (POSIXct, UTC time zone): timestamp of the last price update;

market_cap (double): current market capitalisation;

vol_24h (double): trading volume in the last 24 hours;

price_percent_change_24h (double): percentage change of the price as compared to 24 hours ago.

Details

This function is based on the public CoinGecko API, which has a limit of 50 calls per minute. Please keep this limit in mind when developing your applications.

If no data can be retrieved (e.g. because of a misspecified query parameter), nothing (NULL) will be returned.

Examples

# \donttest{
r <- current_price(
  coin_ids = c("aave", "tron", "bitcoin"),
  vs_currencies = c("usd", "eur", "gbp")
)
print(r)

#> # A tibble: 9 x 7
#>   coin_id      price vs_currency    market_cap    vol_24h price_percent_change_…
#>   <chr>        <dbl> <chr>               <dbl>      <dbl>                  <dbl>
#> 1 aave      311.     usd           3995407541.    3.82e 8                 5.20  
#> 2 aave      263.     eur           3371181048.    3.22e 8                 5.43  
#> 3 aave      225.     gbp           2884728194.    2.76e 8                 5.60  
#> 4 bitcoin 33644      usd         631008558036.    2.00e10                -0.582 
#> 5 bitcoin 28387      eur         532422304963.    1.68e10                -0.370 
#> 6 bitcoin 24291      gbp         455595119996.    1.44e10                -0.207 
#> 7 tron        0.0615 usd           4407717790.    7.50e 8                -0.366 
#> 8 tron        0.0519 eur           3719073594.    6.33e 8                -0.153 
#> 9 tron        0.0444 gbp           3182420730.    5.42e 8                 0.0105
#> # … with 1 more variable: last_updated_at <dttm>

# }