Get historical market data

Retrieves market data for a coin for the last n days

coin_history(coin_id, vs_currency, days, interval = NULL, max_attempts = 3)

Arguments

coin_id
(character): ID of the coin of interest. An up-to-date list of supported coins and their IDs can be retrieved with the supported_coins() function.

vs_currency
(character): name of the base currency to express the price in. An up-to-date list of supported currencies (both fiat and cryptocurrencies) can be obtained with the supported_currencies() function. If an unsupported vs_currency is requested, the call will fail with the respective error message.

days
(numeric or "max"): number of days to look back. If days = "max", the entire available history for coin_id will be retrieved. Depending on the value of days, the time interval used to present the data will differ - see "Details".

interval
(character or NULL): time interval used to present the data. The only currently supported interval is daily. Defaults to NULL.

max_attempts
(double, positive): specifies the maximum number of attempts to call the CoinGecko API (e.g., if the first call fails for some reason). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with the following columns:

timestamp (POSIXct);

coin_id (character): coin ID;

vs_currency (character): same as the argument vs_currency;

price (double): coin price, as of timestamp;

total_volume (double): a 24 h rolling-window trading volume, as of timestamp;

market_cap (double): market capitalisation, as of timestamp.

Details

If days = 1 and interval = NULL, the data will be returned for every few minutes (typically 3-8 minutes). If days is between 2 and 90 (inclusive) and interval = NULL, an (approximately) hourly time step will be used. Daily data are used for days above 90. If interval = "daily", daily data will be used irrespective of the value of days.

This function is based on the public CoinGecko API, which has a limit of 50 calls per minute. Please keep this limit in mind when developing your applications.

Examples

# \donttest{
r <- coin_history(coin_id = "bitcoin", vs_currency = "usd", days = 30)
print(r)

#> # A tibble: 721 x 6
#>    timestamp           coin_id vs_currency  price total_volume    market_cap
#>    <dttm>              <chr>   <chr>        <dbl>        <dbl>         <dbl>
#>  1 2021-06-29 12:23:38 bitcoin usd         35729. 33384839884. 669156220940.
#>  2 2021-06-29 12:02:35 bitcoin usd         35669. 33095224117. 668603462428.
#>  3 2021-06-29 11:03:04 bitcoin usd         35750. 32573828247. 670116373487.
#>  4 2021-06-29 10:01:18 bitcoin usd         35305. 31819761493. 661780725343.
#>  5 2021-06-29 09:01:30 bitcoin usd         35174. 32446664092. 659294360610.
#>  6 2021-06-29 08:00:28 bitcoin usd         34980. 32488822315. 655682920427.
#>  7 2021-06-29 07:01:32 bitcoin usd         35270. 32907363767. 660693065036.
#>  8 2021-06-29 06:02:13 bitcoin usd         34776. 32237919728. 651457474451.
#>  9 2021-06-29 05:01:05 bitcoin usd         34981. 32393707236. 655702350600.
#> 10 2021-06-29 04:02:59 bitcoin usd         34647. 31922071332. 649433960604.
#> # … with 711 more rows

# }