Get historical market data

Retrieves coin-specific market data for the last n days

coin_history(
  coin_id,
  vs_currency = "usd",
  days,
  interval = NULL,
  max_attempts = 3
)

Arguments

coin_id
(character): ID of the coin of interest or a vector with several IDs. The maximum number of coins that can be queried simultaneously is 30. An up-to-date list of supported coins and their IDs can be retrieved with the supported_coins() function.

vs_currency
(character): name of the reference currency to express the price in. An up-to-date list of supported reference currencies (both fiat and cryptocurrencies) can be obtained with the supported_currencies() function. If an unsupported vs_currency is requested, the call will fail with the respective error message.

days
(numeric or "max"): number of days to look back. If days = "max", the entire available history for coin_id will be retrieved. Depending on the value of days, the time interval used to present the data will differ - see "Details".

interval
(character or NULL): time interval used to present the data. The only currently supported value is daily. Defaults to NULL.

max_attempts
(double, positive): specifies the maximum number of attempts to call the CoinGecko API (e.g., if the first call fails for some reason). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with the following columns:

timestamp (POSIXct);

coin_id (character): same as the argument coin_id;

vs_currency (character): same as the argument vs_currency;

price (double): coin price, as of timestamp;

total_volume (double): a 24 hours rolling-window trading volume, as of timestamp;

market_cap (double): market capitalisation, as of timestamp.

Details

If days = 1 and interval = NULL, the data will be returned for every few minutes (typically 3-8 minutes). If days is between 2 and 90 (inclusive) and interval = NULL, an (approximately) hourly time step will be used. Daily data are used for days above 90. If interval = "daily", daily data will be used irrespective of the value of days.

Sometimes, the retrieved data will contain missing values. In such cases, the function will issue a warning and show a list of columns that have missing values.

This function is based on the public CoinGecko API, which has a limit of 50 calls per minute. Please keep this limit in mind when developing your applications.

Examples

# \donttest{
r <- coin_history(coin_id = "bitcoin", vs_currency = "usd", days = 30)
print(r)
#> # A tibble: 722 × 6
#>    timestamp           coin_id vs_currency  price total_volume    market_cap
#>    <dttm>              <chr>   <chr>        <dbl>        <dbl>         <dbl>
#>  1 2021-10-02 16:04:12 bitcoin usd         47997. 31272740081. 903915964900.
#>  2 2021-10-02 17:04:36 bitcoin usd         48102. 31045850229. 905891483117.
#>  3 2021-10-02 18:03:09 bitcoin usd         47967. 30165400568. 903345291339.
#>  4 2021-10-02 19:01:04 bitcoin usd         47963. 30334299295. 902811928038.
#>  5 2021-10-02 20:01:58 bitcoin usd         48230. 30203254702. 908314195739.
#>  6 2021-10-02 21:01:19 bitcoin usd         48160. 30208052696. 906394776221.
#>  7 2021-10-02 22:06:46 bitcoin usd         48272. 30080007505. 908665202698.
#>  8 2021-10-02 23:03:06 bitcoin usd         48223. 29259912919. 908172896162.
#>  9 2021-10-03 00:03:25 bitcoin usd         47863. 29834752196. 901784970578.
#> 10 2021-10-03 01:04:33 bitcoin usd         47361. 29727152861. 901784970578.
#> # … with 712 more rows
# }