Historical OHLC price data

Retrieves open-high-low-close price data for the last n days

coin_history_ohlc(coin_id, vs_currency = "usd", days, max_attempts = 3)

Arguments

coin_id
(character): ID of the coin of interest or a vector with several IDs. The maximum number of coins that can be queried simultaneously is 30. An up-to-date list of supported coins and their IDs can be retrieved with the supported_coins() function.

vs_currency
(character): name of the reference currency to express the price in. An up-to-date list of supported reference currencies (both fiat and cryptocurrencies) can be obtained with the supported_currencies() function. If an unsupported vs_currency is requested, the call will fail with the respective error message.

days
(numeric or "max"): number of days to look back. The only acceptable values are 1, 7, 14, 30, 90, 180, 365 and "max". Attempts to assign any other values will fail with the corresponding error message. If days = "max", the entire available history will be retrieved. Depending on the value of days, the time interval used to present the data will differ - see "Details".

max_attempts
(double, positive): specifies the maximum number of attempts to call the CoinGecko API (e.g., if the first call fails for some reason). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with the following columns:

timestamp (POSIXct): beginning of the candle's time interval;

coin_id (character): same as the argument coin_id;

vs_currency (character): same as the argument vs_currency;

price_open (double): coin price in the beginning of a time interval;

price_high (double): highest coin price observed within a time interval;

price_low (double): lowest coin price observed within a time interval;

price_close (double): coin price in the end of a time interval.

Details

Granularity of the retrieved data (i.e. candle's body) depends on the value of days as follows:

1 day: 30 minutes;

7 - 30 days: 4 hours;

above 30: 4 days.

This function is based on the public CoinGecko API, which has a limit of 50 calls per minute. Please keep this limit in mind when developing your applications.

Examples

# \donttest{
r <- coin_history_ohlc(
  coin_id = "cardano",
  vs_currency = "usd",
  days = 7
)
print(r)
#> # A tibble: 42 × 7
#>    timestamp           coin_id vs_currency price_open price_high price_low
#>    <dttm>              <chr>   <chr>            <dbl>      <dbl>     <dbl>
#>  1 2021-10-25 20:00:00 cardano usd               2.17       2.18      2.17
#>  2 2021-10-26 00:00:00 cardano usd               2.16       2.16      2.15
#>  3 2021-10-26 04:00:00 cardano usd               2.14       2.18      2.14
#>  4 2021-10-26 08:00:00 cardano usd               2.19       2.19      2.18
#>  5 2021-10-26 12:00:00 cardano usd               2.2        2.2       2.19
#>  6 2021-10-26 16:00:00 cardano usd               2.18       2.18      2.16
#>  7 2021-10-26 20:00:00 cardano usd               2.17       2.18      2.15
#>  8 2021-10-27 00:00:00 cardano usd               2.17       2.19      2.14
#>  9 2021-10-27 04:00:00 cardano usd               2.15       2.15      2.15
#> 10 2021-10-27 08:00:00 cardano usd               2.15       2.16      2.14
#> # … with 32 more rows, and 1 more variable: price_close <dbl>
# }