History snapshot for a coin

Coin-specific market data for a given historical date

coin_history_snapshot(coin_id, date, vs_currencies = "usd", max_attempts = 3)

Arguments

coin_id
(character): ID of the coin of interest. An up-to-date list of supported coins and their IDs can be retrieved with the supported_coins() function.

date
(Date): date of interest. If no data exist for the requested date, nothing (NULL) will be returned.

vs_currencies
(character): a vector with names of the reference currencies to express the price in, e.g. c("usd", "eur", "btc"). An up-to-date list of supported vs_currencies (both fiat and cryptocurrencies) can be obtained with the supported_currencies() function. If vs_currencies contains at least one unsupported currency, the call will fail with the respective error message.

max_attempts
(double, positive): specifies the maximum number of attempts to call the CoinGecko API (e.g., if the first call fails for some reason). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

If the requested data exist, this function will return a tibble with with as many rows as the length of vs_currencies and the following columns:

coin_id (character): same as the argument coin_id;

symbol (character): symbol of the coin;

name (character): common name of the coin;

date (Date): same as the argument date;

vs_currency (character): reference currency, in which the price is expressed (ordered alphabetically);

price (double): price of the coin;

market_cap (double): market capitalisation of the coin;

total_volume (double): total trading volume recorded on that date.

Details

This function is based on the public CoinGecko API, which has a limit of 50 calls per minute. Please keep this limit in mind when developing your applications.

Examples

# \donttest{
r <- coin_history_snapshot(
  coin_id = "cardano",
  date = as.Date("2021-05-01"),
  vs_currencies = c("usd", "eth")
)
print(r)

#> # A tibble: 2 x 8
#>   coin_id symbol name    date       vs_currency   price  market_cap total_volume
#>   <chr>   <chr>  <chr>   <date>     <chr>         <dbl>       <dbl>        <dbl>
#> 1 cardano ada    Cardano 2021-05-01 eth         4.88e-4     1.56e 7     1040152.
#> 2 cardano ada    Cardano 2021-05-01 usd         1.35e+0     4.33e10  2888334739.

# }